algo_seed/requirements.txt
psycopg2-binary
gspread
google-auth
python-dateutil
